requirements.txt
streamlit==1.39.0
docx2txt==0.8
PyPDF2==3.0.1
google-generativeai==0.8.3

